Return the actual paths clicked in the clicks API
Instead of just returning the position and the number of clicks, also
return a list of [path, number of clicks] for each path that has had
results clicked on at that position.  When search results move around,
this is essential so that we know which search result actually got the
clicks.